system, launcher, controller: modify to build the box; launcher_lib: create new file; examples/alanine_dipeptide/test_ala_build_box: create new file
