+++ adaptivemastermsm/launcher/launcher.py
@@ -11,74 +11,65 @@ import multiprocessing as mp
 
 # AdaptiveMasterMSM
 from adaptivemastermsm.system import system
+from ..launcher import launcher_lib
 
 class Launcher(object):
     """
     Call GROMACS and process output to pass it to class analyzer
     """
 
-    def __init__(self, md_step, forcefield, water, \
-            filepdb, wet_xtc_file,  dry_xtc_file=None, last_wet_snapshot=None):
+    def __init__(self, md_step, pdb, forcefield, water, \
+            wet_xtc_file,  dry_xtc_file=None, last_wet_snapshot=None):
         """
         Read input from system, run MD, and process output
         for the class analyzer
 
         Parameters
         ----------
-        sysfile : str
-            File containing system information?
         md_step : str
             'Equilibration' or 'Production'
-        water: str
-            water model for GROMACS (tip3p, ...)
-        filepdb : str
-            PDB file defining the system
+        pdb : str
+            Path to the PDB or GRO file to start setup
+        forcefield : str
+            Name of force field in Gromacs format (e.g. amber03, charmm27...)
+        water : str
+            Water model for in Gromacs format (e.g. tip3p...)
         wet_xtc_file : str
-            
+
         dry_xtc_file : str
             
         last_wet_snapshot : str
-            
-        Returns
-        -------
-        trajfile : str
-            File containing processed trajectories
 
         """
-        self.pdb = filepdb
+        self.pdb = pdb
         self.dry_xtc_file = dry_xtc_file
         self.last_wet_snapshot = last_wet_snapshot
 
-        self.params = self.make_system()
-
-        if params.md_step == 'Production':
+        if md_step == 'Production':
             self.wet_xtc_file = wet_xtc_file           
             n_short_runs = 2 # this will be managed by controller
-            mdp_template = params.filemdp
 
-            # ndx = ??? use this to set positions of ligands in future work
+            # ndx ? use this to set positions of ligands in future work
             for i in range(n_short_runs):
-                mdp = self.gen_mdp(mdp_template, i)
+
+                mdp = self.gen_mdp(i)
                 # different resampling positions:
                 gro = "data/gro/%s.gro" % i
-                if not os.path.exists(gro[:gro.rfind("/")]):
-                    os.makedirs(gro[:gro.rfind("/")])
+                launcher_lib.checkfile(gro)
                 cmd = "cp %s data/gro/%s.gro" % (self.pdb, i)
                 os.system(cmd)
                 top = "data/top/%s.top" % i
-                if not os.path.exists(top[:top.rfind("/")]):
-                    os.makedirs(top[:top.rfind("/")])
+                launcher_lib.checkfile(top)
                 itp = "data/itp/%s.itp" % i
-                if not os.path.exists(itp[:itp.rfind("/")]):
-                    os.makedirs(itp[:itp.rfind("/")])
-                cmd = 'gmx pdb2gmx -f %s -o processed_%s -ff %s -water %s' % \
-                    (gro, self.pdb, forcefield, params.water)
+                launcher_lib.checkfile(itp)
+                cmd = 'gmx pdb2gmx -f %s -o processed_%s -ff %s -water %s'%\
+                      (gro, self.pdb, forcefield, water)
                 print(" running: ",cmd)
                 p = subprocess.Popen(shlex.split(cmd), stdout=subprocess.PIPE, stderr=subprocess.PIPE)
                 output, error = p.communicate()
                 #print(output, error)
-                cmd = "mv processed_%s %s & mv topol.top %s & mv posre.itp %s" % \
-                        (self.pdb, gro, top, itp)
+                cmd = "mv processed_%s %s & mv topol.top %s & mv posre.itp %s"%\
+                      (self.pdb, gro, top, itp)
                 os.system(cmd)
                 out = "data/tpr/%s.tpr" % i
                 output, error = self.gen_tpr(gro, top, mdp, out, i)
@@ -98,16 +89,9 @@ class Launcher(object):
                 out, err = result.get()
                 print("out: {} err: {}".format(out, err))
 
-        elif params.md_step == 'Equilibration':
+        elif md_step == 'Equilibration':
             # first equilibrate
-            cmd = 'gmx pdb2gmx -f %s -o processed_%s -ff %s -water %s -ignh' % \
-                    (self.pdb, self.pdb, forcefield, params.water)
-            print(" running: ",cmd)
-            p = subprocess.Popen(shlex.split(cmd), stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            out, err = p.communicate()
-            #print(out, err)
-            out, err = params.build_box(self.pdb, 1)
-            #print(out, err)
+            # build_box, etc
             user_wet_xtc_file = wet_xtc_file
             self.wet_xtc_file = 'equilibration.xtc'
             self.run_md(4, params)
@@ -122,12 +106,12 @@ class Launcher(object):
 
         return
 
-    def make_system(self):
+    def make_system(self, ff, wat):
         """ 
         Creates an instance of System class to define the system and parameters
 
         """
-        params = system.System(self.water, self.md_step, 1)
+        params = system.System(self.pdb, forcefield=ff, water=wat, copy=1)
         return params
 
     def gen_tpr(self, gro, top, mdp, tpr, i):
@@ -135,8 +119,7 @@ class Launcher(object):
         Function for generating tpr files
 
         """
-        if not os.path.exists(tpr[:tpr.rfind("/")]):
-            os.makedirs(tpr[:tpr.rfind("/")])
+        launcher_lib.checkfile(tpr)
         cmd = "gmx grompp -c %s -p %s -f %s -o %s -maxwarn 1"\
                 %(gro, top, mdp, tpr)
         print(" running: ",cmd)
@@ -147,21 +130,22 @@ class Launcher(object):
         os.system(cmd)
         return out, err
     
-    def gen_mdp(self, filemdp, i):
+    def gen_mdp(self, i):
         """ 
         Generate Gromacs parameter input file (mdp)
         
         """
+        txt = launcher_lib.write_mdp()
         filemdpout = "data/mdp/%s.mdp" % i
-        try:
-            out = open(filemdpout, "w")
-        except IOError:
-            os.makedirs(filemdpout[:filemdpout.rfind("/")])
-            out = open(filemdpout, "w")
+        launcher_lib.checkfile(filemdpout)
+        out = open(filemdpout, "w")
+        out.write(txt)
+        out.close()
+        """ keep for a while to consider other options:
         rawmdp = open(filemdp).readlines()
         for i in rawmdp:
             out.write("%s"%i)
-        out.close()
+        """
         return filemdpout
     
     def gromacs_worker(self, x):
@@ -194,7 +178,7 @@ class Launcher(object):
         # overwrite each other, until I see their need
 
         cmd = 'gmx grompp -f %s -c processed_%s -p topol.top -maxwarn 1'\
-                % (params.filemdp, self.pdb); \
+                % (params.filemdp, self.pdb) &\
                 'gmx mdrun -nt %d -s topol.tpr -x %s -c processed_%s -g prod.log' % \
                 (n_threads, self.pdb, self.wet_xtc_file)
         print(" running: ",cmd)
@@ -227,3 +211,5 @@ class Launcher(object):
         os.system('mv equilibration.xtc *.mdp *.gro logs')
     
         return
+
+
